# Add Inhibitor.MutesAll()