@Nuckal777 2024-08-22
>  Do you intend to call it from MuteStage as well as APIv2?

Yes.

> The PR seems incomplete without it?

I thought I check the general interest for the topic before implementing everything. I will add it. Might take a few days.