@Nuckal777 2024-08-27
I added `MutesAll` to the api and `MuteStage` code paths. The CI needs a re-run, network foo during setup.