@Nuckal777 2024-09-04
> Is it possible to optimize inhibition rules without changing the Mutes interface?

I implemented that. In the current state the algorithmic improvement can only be realised when all alerts are passed to `MutesAll()`. By moving `setAlertStatus` out of `alertFilter` the different API handlers can choose how to update it. For `getAlertsHandler` I used `MutesAll()` here.