@grobinson-grafana 2024-09-04
> > Is it possible to optimize inhibition rules without changing the Mutes interface?
> 
> I implemented that. In the current state the algorithmic improvement can only be realised when all alerts are passed to `MutesAll()`.

I looked again at the code, but I'm afraid I don't see where the optimization is implemented for `Mutes()`? Like you said, the performance optimization only works if you pass all alerts to `MutesAll()`, and there is no performance improvement if you only call `Mutes()`? Did I misunderstand?

My suggestion was to see if we can optimize the performance of `Mutes` without having to pass all alerts to `MutesAll().` For example, what if the cache is persistent, like `matcherCache` for silences?