@Nuckal777 2024-09-06
> there is no performance improvement if you only call `Mutes()`?

Exactly. The whole speed improvement comes from caching the source cache evaluation of `InhibitRules` across multiple alerts. In the context of inhibitions, that's the whole set of potential target alerts. For `MutesAll()` it roughly boils down to:

```go
for _, r := range ih.rules {
    var alerts []*types.Alert // <- reused in the inner loop
    var scacheEval []bool // <- reused in the inner loop
    for i, lset := range lsets {
        // ...
        if inhibitedByFP, eq := r.hasEqualCached(lset, r.SourceMatchers.Matches(lset), alerts, scacheEval); eq {
            ih.marker.SetInhibited(fingerprints[i], inhibitedByFP.String())
        }
        // ...
    }
}
```

Just using `Mutes()` flips the two loops. That looks roughly like:

```go
func someOtherFunc() {
    // ...
    for _, a := range alerts {
        inhibitor.Mutes(a.Labels) // <- evaluates the source cache for each a
    }
}

func (ih *Inhibitor) Mutes(lset model.LabelSet) bool {
    // ...
    for _, r := range ih.rules {
        // ...
        if inhibitedByFP, eq := r.hasEqual(lset, r.SourceMatchers.Matches(lset)); eq { // <- always evaluates the source cache matches
            ih.marker.SetInhibited(fp, inhibitedByFP.String())
            return true
        }
    }
}
```

Caching the source cache evaluations "requires" to pass a list of alerts. If the signature change is undesirable, we can close this PR. It doesn't work without. Technically, the state `MutesAll()` caches can be externalized into a struct, which implements the `Muter` interface, e.g.:

```go
type CachingInhibitor {
    base Inhibitor
    var alerts [][]*types.Alert // one cache for each inhibition rule
    var scacheEval [][]bool // one cache for each inhibition rule
}
```

Externalizing the state would cost more memory, likely requires locking and requires explicitly dropping these caches at some point (likely after calling `Mutes()` a few times).

> My suggestion was to see if we can optimize the performance of Mutes without having to pass all alerts to MutesAll(). For example, what if the cache is persistent, like matcherCache for silences?

I expect something in that direction is possible as well, but it's something for different PR.