@Spaceman1701 2025-11-12
Hi! I know it's been a while, is this change something you're still interested in working on? 

Inhibitor performance has also been improved a lot, so I'm not sure what the performance gains of this PR are anymore. If you still think it'd be useful (and have the time), the first step is probably to rebase and collect new benchmarks.