@Nuckal777 2025-11-21
Hey, thanks for the information. We tried the latest release, but haven't seen a significant improvement unfortunately.

> If you still think it'd be useful (and have the time), the first step is probably to rebase and collect new benchmarks.

I'll have to see, if I get around to it. 😢