@SoloJacobs 2026-08-20
Hi @Nuckal777 ,

This PR is mostly outdated. We have since landed four improvements to the inhibitor performance. Some of those ideas are similar to your improvements:

* #4607
* #4668
* #5020

Thanks for pushing on this. Your batching idea is a good one and the benchmark cases you added for multiple target alerts cover ground we still don't, so if you'd like to open a smaller PR with just those I'd happily review it.

But for now, I'll close this PR.

Kind regards